S01E03_dynamic_answering/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.12.0
python-dotenv==1.0.1 